Fix subprocess.Popen() execution.
As outlined in https://docs.python.org/3/library/subprocess.html#subprocess.Popen -
"args should be a sequence of program arguments or else a single string. By default, the program to execute is the first item in args if args is a sequence. If args is a string, the interpretation is platform-dependent."

This commit turns args into a sequence, which makes the plugin work correctly on POSIX systems (Linux/OSX).