Fix Addon Using Legacy Pose Library
There is also a filter to prevent the addon from choosing poses that have bones that aren't in the current model. Also, I told it to use the pose assets to key frame the model and it stores pose assets in the mouth options. The addon also will only key the bones keyed in the pose assets for the model that were chosen.